03_app/requirements.txt
PyYAML==3.11
docopt==0.6.2
pymongo==2.8
six==1.5.2
celery==3.1.17
